requirements.txt
dj-database-url==0.5.0
Django==2.1.15
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
python-decouple==3.3
pytz==2020.1
whitenoise==5.1.0
psycopg2==2.7
